requirements.txt
arch==5.3.1
image_quality==1.2.7
matplotlib==3.5.3
numpy
opencv_contrib_python==4.7.0.72
Pillow==9.5.0
scipy==1.7.3
thop
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
einops==0.6.1
lpips==0.1.4
opencv-python
gradio
safetensors